- More reorg/module splitting of original model_select.py. Moved model interfaces (sklearn/xgb wrappers) and averaging regressor out to their own files.
