+++ pymlkit/models/ifaces/model_select.py
@@ -20,22 +20,16 @@ from sklearn.svm import SVC
 # Regressor import
 from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor
 from sklearn.ensemble import GradientBoostingRegressor, AdaBoostRegressor
-from sklearn.linear_model import LinearRegression, SGDRegressor
-from sklearn.neural_network import MLPRegressor
+from sklearn.linear_model import LinearRegression
 from sklearn.svm import SVR
 # Misc ML import
-from sklearn.base import BaseEstimator, RegressorMixin
-import xgboost as xgb
 
-from sklearn.externals import joblib
-
-
-__all__ = ['clf_scan',
-           'reg_scan',
-           'reg_scoring',
-           'make_majvote',
-           'AveragingRegressor',
-           'save_model']
+__all__ = [
+    'clf_scan',
+    'reg_scan',
+    'reg_scoring',
+    'make_majvote'
+]
 
 
 # ====== Model Comparisons/scans
@@ -202,191 +196,3 @@ def make_majvote(estimators=None, estimator_names=None, vote_meth='hard'):
                               voting=vote_meth)
 
     return vc
-
-
-class AveragingRegressor(BaseEstimator, RegressorMixin):
-    # TODO: Modify how parameters are used. Put them all into a dict. Also change X_train, y_train to just X,y
-    """
-    Summary: A Meta-regressor that averages all predictions of it's consituent regressors. Analogous to
-        a majority vote classifer, but for regressoion
-
-    Attributes:
-    -------------
-        - regs: Base/Constituent regressors from which the average predictions are calculated
-        - reg_names: Names of the constituent regressors
-        - params: Optionally user-supplied initialization parameters for the
-        - base_predictions: Predictions of the constituent classifiers. This attribute is None until the predict method
-                is called
-        - avg_predictions: Average predictions calculated from the predictions of the constituent regressors.
-    """
-    def __init__(self, regressors=None, regressor_names=None, init_parameters=None, verbose=0):
-        """
-        Initialization
-        :param regressors: (obj list) - Constituent regressors of AveragingRegressor
-        :param regressor_names: (str list) - Names of the constituent regressors
-        :param init_parameters: (dict list) - initialization parameters for the corresponding regressors. These
-                must be passed as a list of dictionaries s.t. the parameters in each index are the corresponding
-                paramters for the regressor at the same index in the 'regressors' parameter. Can provide a partial
-                list, containing parameter dictionaries only for the first few regressors.
-        """
-        self.params = {'regressors':      regressors,
-                       'regressor_names': regressor_names,
-                       'init_parameters': init_parameters,
-                       'verbose': verbose}
-        self.regs = regressors
-        self.reg_names = regressor_names
-        self.reg_params = init_parameters
-        self.verbose = verbose
-        self.base_predictions = None
-        self.avg_predictions = None
-        super().__init__()
-        super().set_params(**self.params)
-
-        # Return error if no constituent regressors are supplied
-        if regressors is None:
-            raise TypeError("Parameter 'regressors' should be a list of estimators with base scikit-learn regressor"
-                            " methods.")
-
-        # Initialize constituent regressors with custom parameters if they are provided
-        if init_parameters is not None:
-            for i in range(len(self.reg_params)):
-                self.regs[i] = self.regs[i](**self.reg_params[i])
-
-    def fit(self, X_train, y_train=None):
-        """
-        Method to fit all Regressors
-        :param X_train: (pandas df) - Training features
-        :param y_train: (pandas series) - Training target variable
-        :return: None
-        """
-        print('=> Fitting AveragingRegressor:')
-        for i in range(len(self.regs)):
-            if self.verbose > 0:
-                print('==> Fitting %s' % self.reg_names[i])
-            self.regs[i].fit(X_train, y_train)
-
-    def predict(self, X_test):
-        """
-        Method to predict target variable values. Final results are the average of all predictions
-        :param X_test: (pandas df) - Test features
-        :return: self.avg_predictions: (np.array) Average target variable predictions
-        """
-        predictions = {}
-        average_predictions = np.zeros(shape=(len(X_test)), dtype=np.float64)
-
-        if len(self.reg_names) == len(self.regs):
-            add_names = True
-        for i in range(len(self.regs)):
-            y_pred = self.regs[i].predict(X_test)
-            average_predictions += y_pred
-            name = self.reg_names[i] if add_names else ('Regressor%i' % i)
-            predictions.setdefault(name, y_pred)
-
-        average_predictions /= float(len(self.regs))
-        predictions.setdefault('Average', average_predictions)
-        self.base_predictions = predictions
-        self.avg_predictions = average_predictions
-
-        return self.avg_predictions
-
-
-# ====== Wrappers for ensuring methods consistency between ML packages
-class XgbWrapper(object):
-    """
-    Summary: Wrapper for estimator objects from the XGBoost package. Wraps class to provide access to
-        uniform fit/train/predict methods. Credit for original code goes to user Eliot
-        Barril on kaggle.com.
-
-    Attributes:
-    -------------
-        - gbdt: Gradient boosting decision tree from xgboost package. Created upon fit method invocation.
-        - params: Estimator parameters for gradient boosting decision tree
-        - nrounds: # of rounds parameter for XGboost estimator
-    """
-    def __init__(self, seed=0, params=None):
-        """
-        :param seed: (int) - Random seed
-        :param params: (dict) - Estimator parameters
-        """
-        self.gbdt = None
-        self.params = params
-        self.params['seed'] = seed
-        self.nrounds = params.pop('nrounds', 250)   # pop nrounds from dict w/ 250 as default if not passed
-
-    def train(self, X_train, y_train):
-        """
-        Method to train a gradient boosting decision tree from the xgboost package, formatted for a generic 'train'
-            interface.
-        :param X_train: Training features
-        :param y_train: Training target/labels
-        :return: None
-        """
-        dtrain = xgb.DMatrix(X_train, y_train)
-        self.gbdt = xgb.train(self.params, dtrain, self.nrounds)
-
-    def fit(self, X_train, y_train):
-        """
-        Method to train a gradient boosting decision tree from the xgboost package, formatted for a
-            sklearn-like interface
-        :param X_train: Training features
-        :param y_train: Training target/labels
-        :return: None
-        """
-        dtrain = xgb.DMatrix(X_train, y_train)
-        self.gbdt = xgb.train(self.params, dtrain, self.nrounds)
-
-    def predict(self, X):
-        """
-        Method calculate estimator predictions
-        :param X: Feature set from which to predict targets/labels
-        :return: predicted values.
-        """
-        return self.gbdt.predict(xgb.DMatrix(X))
-
-
-class SklearnWrapper(object):
-    """
-    Summary: Wrapper for estimator objects from the XGBoost package. Wraps class to provide access to
-        uniform fit/train/predict methods. Credit for original code goes to user Eliot Barril and posted to
-        kaggle.com for "House Prices: Advanced Regression Techniques."
-
-    Attributes
-    ------------
-        - est: Sci-kit learn estimator passed during initialization
-        - params: Parameters to pass to the scikit-learn estimator
-    """
-    def __init__(self, est, seed=0, params=None):
-        """
-        :param est: (scikit-learn estimator) - Estimator to be wrapped
-        :param seed: (int) - Random seed
-        :param params: (dict) - Estimator parameters
-        """
-        self.params = params
-        self.params['random_state'] = seed
-        self.est = est(**self.params)  # pass dict to
-
-    def train(self, x_train, y_train):
-        """
-        Method to train estimator package, formatted for a generic 'train' interface.
-        :param X_train: Training features
-        :param y_train: Training target/labels
-        :return: None
-        """
-        self.est.fit(x_train, y_train)
-
-    def fit(self, x_train, y_train):
-        """
-        Method to train estimator formatted for a sklearn-like interface
-        :param X_train: Training features
-        :param y_train: Training target/labels
-        :return: None
-        """
-        self.est.fit(x_train, y_train)
-
-    def predict(self, x):
-        """
-        Method calculate estimator predictions
-        :param X: Feature set from which to predict targets/labels
-        :return: predicted values.
-        """
-        return self.est.predict(x)
